==> notebooks/00_intro.ipynb <==
**Contexto**

A Rossmann opera mais de 3.000 farmácias em 7 países europeus. Atualmente, os gerentes de loja têm a tarefa de prever suas vendas diárias com até seis semanas de antecedência. As vendas são influenciadas por diversos fatores, incluindo promoções, concorrência, feriados escolares e estaduais, sazonalidade e localidade. Com milhares de gerentes prevendo vendas com base em suas circunstâncias específicas, a precisão dos resultados pode ser bastante variável.

**Descrição do dataset**

Dados históricos de vendas de 1.115 lojas Rossmann. A tarefa é prever a coluna "Sales" para o conjunto de teste. Observe que algumas lojas no conjunto de dados foram temporariamente fechadas para reforma.

**Arquivos**
- train.csv - Dados históricos de vendas.
- production.csv - Dados históricos expluindo vendas para as próximas 6 semanas.
- store.csv - Dados suplementares para as lojas.

**Variáveis**

- **Id** - Um número de identificação para cada para par (Store, Date).
- **Store** - Identificador único de loja.
- **Sales** - O faturamento de um determinado dia (variável alvo)
- **Customers** - O número de clientes em um determinado dia.
- **Open** - Um indicador para saber se a loja estava aberta: *0 = fechada*, *1 = aberta*.
- **StateHoliday** - Indica um feriado estadual. Normalmente, todas as lojas, com poucas exceções, fecham em feriados estaduais. Todas as escolas fecham em feriados e fins de semana. *a = feriado, b = Páscoa, c = Natal, 0 = Nenhum*.
- **SchoolHoliday** - Indica se a (Loja, Data) foi afetada pelo fechamento das escolas públicas.
- **StoreType** - Diferencia entre 4 modelos de loja diferentes: *a, b, c, d*
- **Assortment** - Descreve um nível de variedade: *a = básico, b = extra, c = estendido*
- **CompetitionDistance** - Distância em metros até a loja concorrente mais próxima.
- **CompetitionOpenSince[Month/Year]** - Fornece o ano e o mês aproximados da abertura da loja concorrente mais próxima
- **Promo** - Indica se uma loja está realizando uma promoção naquele dia.
- **Promo2** - Indica uma promoção contínua e consecutiva para algumas lojas: *0 = a loja não está participando, 1 = a loja está participando*
- **Promo2Since[Year/Week]** - Descreve o ano e a semana do calendário em que a loja começou a participar da Promo2.
- **PromoInterval** - Descreve os intervalos consecutivos em que a Promoção2 está iniciado, indicando os meses em que a promoção é reiniciada. Por exemplo, "Feb,May,Aug,Nov" significa que cada rodada começa em fevereiro, maio, agosto, novembro de qualquer ano para aquela loja.

---

**Premissas adotadas para tratar dos dados vazios:**

- Considera-se que não há concorrência nas proximidades de lojas cujas que linhas que não possuem indicação de distância do concorrente ou seu tempo de existência, portanto os vazios serão preenchidos com um número muito superior a distância máxima no banco de dados para casos de dados vazios na distância e, que a data de abertura é igual a data presente, totalizando um período de existência de zero meses/anos.

- Considera-se que não há promoção ocorrendo nas lojas cujas linhas não possuem indicação de tempo de início da promoção, portanto os vazios serão preenchidos data igual a data presente, totalizando um período de promoção de zero semanas/anos.

- Para tratar dos dados vazios na variável 'promo_interval', ela foi convertida em uma variável chamada 'is_promo' que indica se a data atual está contida dentro do intervalo de promoção (1) ou não (0).

**Premissas adotadas na filtragem de dados:**

- Apenas dias em que as lojas ficaram abertas serão estudados.
