==> notebooks/03_filtering_to_business.ipynb <==
# Bibliotecas

import warnings

import pandas       as pd

warnings.filterwarnings( 'ignore' )

# Importação de dados

df_02 = pd.read_pickle('../exports/cicle_exports/02_feature_eng/df_02.pkl', compression="gzip")

df_03 = df_02.copy()

# 3.0 Filtragem de variáveis para o negócio

## 3.1. Filtragem de linhas

# Filtrando linhas apenas para dias em que houve vendas:
df_03 = df_03[(df_03['open'] != 0) & (df_03['sales'] > 0)]

## 3.2. Seleção de colunas

# Removendo colunas:
cols_drop = ['customers', 'open', 'competition_open_since_year', 'competition_open_since_month', 'promo2_since_year', 'promo2_since_week', 'promo_interval', 'month_map']
df_03 = df_03.drop( cols_drop, axis=1 )

**Motivo de remoção:**

- **'customers'**: O número de clientes que vão às lojas não pode ser obtido de antemão, portanto não é útil como feature para um algoritmo para prever as vendas. Essa é praticamente uma segunda variável alvo.

- **'open'**: Como foi realizado um filtro para considerar apenas os dias em que as lojas ficam abertas, essa variável torna-se irrelevante.

- **'competition_open_since_year' e 'competition_open_since_month'**: Foram usadas como referência para criar outras variáveis e não possuem mais utilidade prática.

- **'promo2_since_year' e 'promo2_since_week'**: Foram usadas como referência para criar outras variáveis e não possuem mais utilidade prática.

- **'promo_interval' e 'month_map'**: Foram usadas como referência para criar outras variáveis e não possuem mais utilidade prática.

df_03.head()

# Exportação de resultados

df_03.to_pickle('../exports/cicle_exports/03_filtering_to_business/df_03.pkl', compression='gzip')
